pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "customer-support-triage-openenv"
version = "1.0.0"
description = "A real-world environment simulating a customer support agent triaging, categorizing, and prioritizing incoming tickets."
authors = [
    {name = "Kartikey Singh", email = "kartikey@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["openenv", "reinforcement-learning", "customer-support", "ai", "agent-evaluation"]
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "openenv>=0.1.0",
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/kartikey-singh-k/customer-support-openenv"
Repository = "https://github.com/kartikey-singh-k/customer-support-openenv"
"Bug Tracker" = "https://github.com/kartikey-singh-k/customer-support-openenv/issues"

[project.scripts]
server = "server.app:main"
customer-support-server = "server.app:main"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-dir]
src = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# OpenEnv specific configuration
[tool.openenv]
name = "customer-support-triage"
version = "1.0.0"
entrypoint = "src.env:CustomerSupportEnv"
server_port = 7860
tasks = ["easy", "medium", "hard"]

[tool.openenv.action_space]
ticket_id = "str"
category = ["Billing", "Tech Support", "Refund", "General"]
priority = ["Low", "Medium", "High", "Urgent"]

[tool.openenv.observation_space]
open_tickets = "List[Ticket]"
resolved_tickets = "List[Ticket]"
last_feedback = "str"
current_step = "int"